+++ core/src/Cabana_NeighborList.hpp
@@ -72,165 +72,6 @@ class NeighborList
 };
 
 //---------------------------------------------------------------------------//
-// Neighbor Parallel For
-//---------------------------------------------------------------------------//
-// Algorithm tags.
-
-//! Neighbor operations are executed in serial on each particle thread.
-class SerialNeighborOpTag {};
-
-//! Neighbor operations are executed in parallel in a team on each particle
-//! thread.
-class TeamNeighborOpTag {};
-
-//---------------------------------------------------------------------------//
-/*!
-  \brief Execute \c functor in parallel according to the execution \c policy
-  with a thread-local serial loop over particle neighbors.
-
-  \tparam FunctorType The functor type to execute.
-
-  \tparam NeighborListType The neighbor list type.
-
-  \tparam ExecParams The Kokkos range policy parameters.
-
-  \param exec_policy The policy over which to execute the functor.
-
-  \param functor The functor to execute in parallel
-
-  \param list The neighbor list over which to execute the neighbor operations.
-
-  \param tag Algorithm tag indicating a serial loop strategy over particle
-  neighbors.
-
-  \param str An optional name for the functor. Will be forwarded to the
-  Kokkos::parallel_for called by this code and can be used for identification
-  and profiling purposes.
-
-  A "functor" is a class containing the function to execute in parallel, data
-  needed for that execution, and an optional \c execution_space typedef.  Here
-  is an example functor for neighbor parallel_for:
-
-  \code
-  class FunctorType {
-  public:
-  typedef  ...  execution_space ;
-  void operator() ( const int particle_index, const int neighbor_index ) const ;
-  };
-  \endcode
-
-  In the above example, \c Index is a Cabana index to a given AoSoA element
-  for a particle and its neighbor. Its <tt>operator()</tt> method defines the
-  operation to parallelize, over the range of indices
-  <tt>idx=[begin,end]</tt>.  This compares to a single iteration \c idx of a
-  \c for loop.
-*/
-template<class FunctorType, class NeighborListType, class ... ExecParameters>
-inline void neighbor_parallel_for(
-    const Kokkos::RangePolicy<ExecParameters...>& exec_policy,
-    const FunctorType& functor,
-    const NeighborListType& list,
-    const SerialNeighborOpTag& tag,
-    const std::string& str = "" )
-{
-    std::ignore = tag;
-
-    Kokkos::parallel_for(
-        exec_policy,
-        KOKKOS_LAMBDA( const int i )
-        {
-            for ( int n = 0;
-                  n < NeighborList<NeighborListType>::numNeighbor(list,i);
-                  ++n )
-                functor(
-                    i, NeighborList<NeighborListType>::getNeighbor(list,i,n) );
-        },
-        str );
-
-    Kokkos::fence();
-}
-
-//---------------------------------------------------------------------------//
-/*!
-  \brief Execute \c functor in parallel according to the execution \c policy
-  with team parallelism over particle neighbors.
-
-  \tparam ExecutionSpace The execution space in which to execute the functor.
-
-  \tparam FunctorType The functor type to execute.
-
-  \tparam NeighborListType The neighbor list type.
-
-  \param exec_policy The policy over which to execute the functor.
-
-  \param functor The functor to execute in parallel
-
-  \param list The neighbor list over which to execute the neighbor operations.
-
-  \param tag Algorithm tag indicating a team parallel strategy over particle
-  neighbors.
-
-  \param str An optional name for the functor. Will be forwarded to the
-  Kokkos::parallel_for called by this code and can be used for identification
-  and profiling purposes.
-
-  A "functor" is a class containing the function to execute in parallel, data
-  needed for that execution, and an optional \c execution_space typedef.  Here
-  is an example functor for neighbor parallel_for:
-
-  \code
-  class FunctorType {
-  public:
-  typedef  ...  execution_space ;
-  void operator() ( const int particle_index, const int neighbor_index ) const ;
-  };
-  \endcode
-
-  In the above example, \c Index is a Cabana index to a given AoSoA element
-  for a particle and its neighbor. Its <tt>operator()</tt> method defines the
-  operation to parallelize, over the range of indices
-  <tt>idx=[begin,end]</tt>.  This compares to a single iteration \c idx of a
-  \c for loop.
-*/
-template<class FunctorType, class NeighborListType, class ... ExecParameters>
-inline void neighbor_parallel_for(
-    const Kokkos::RangePolicy<ExecParameters...>& exec_policy,
-    const FunctorType& functor,
-    const NeighborListType& list,
-    const TeamNeighborOpTag& tag,
-    const std::string& str = "" )
-{
-    std::ignore = tag;
-
-    using ExecutionSpace =
-        typename Kokkos::RangePolicy<ExecParameters...>::execution_space;
-
-    using kokkos_policy =
-        Kokkos::TeamPolicy<ExecutionSpace,
-                           Kokkos::IndexType<int>,
-                           Kokkos::Schedule<Kokkos::Dynamic> >;
-    kokkos_policy team_policy( exec_policy.end() - exec_policy.begin(),
-                            Kokkos::AUTO );
-
-    Kokkos::parallel_for(
-        team_policy,
-        KOKKOS_LAMBDA( const typename kokkos_policy::member_type& team )
-        {
-            auto i = team.league_rank();
-            auto num_n = NeighborList<NeighborListType>::numNeighbor( list, i );
-            Kokkos::parallel_for(
-                Kokkos::TeamThreadRange(team,num_n),
-                [&]( const int n ) {
-                    functor(
-                        i, NeighborList<NeighborListType>::getNeighbor(list,i,n) );
-                });
-        },
-        str );
-
-    Kokkos::fence();
-}
-
-//---------------------------------------------------------------------------//
 
 } // end namespace Cabana
