+++ core/src/Cabana_Parallel.hpp
@@ -13,6 +13,7 @@
 #define CABANA_PARALLEL_HPP
 
 #include <Cabana_ExecutionPolicy.hpp>
+#include <Cabana_NeighborList.hpp>
 
 #include <Kokkos_Core.hpp>
 
@@ -29,7 +30,7 @@ namespace Impl
 template<class WorkTag, class FunctorType>
 KOKKOS_FORCEINLINE_FUNCTION
 typename std::enable_if<std::is_same<WorkTag,void>::value>::type
-simdFunctorDispatch( const FunctorType& functor,
+functorTagDispatch( const FunctorType& functor,
                      const std::size_t s,
                      const int a )
 {
@@ -40,7 +41,7 @@ simdFunctorDispatch( const FunctorType& functor,
 template<class WorkTag, class FunctorType>
 KOKKOS_FORCEINLINE_FUNCTION
 typename std::enable_if<!std::is_same<WorkTag,void>::value>::type
-simdFunctorDispatch( const FunctorType& functor,
+functorTagDispatch( const FunctorType& functor,
                      const std::size_t s,
                      const int a )
 {
@@ -48,8 +49,10 @@ simdFunctorDispatch( const FunctorType& functor,
     functor(t,s,a);
 }
 
-}
+} // end namespace Impl
 
+//---------------------------------------------------------------------------//
+// SIMD Parallel For
 //---------------------------------------------------------------------------//
 /*!
   \brief Execute a vectorized functor in parallel with a 2d execution policy.
@@ -114,8 +117,177 @@ inline void simd_parallel_for(
                                            exec_policy.arrayBegin(s),
                                            exec_policy.arrayEnd(s)),
                 [&]( const int a )
-                { Impl::simdFunctorDispatch<work_tag>(functor,s,a);});
+                { Impl::functorTagDispatch<work_tag>(functor,s,a);});
         });
+
+    Kokkos::fence();
+}
+
+
+//---------------------------------------------------------------------------//
+// Neighbor Parallel For
+//---------------------------------------------------------------------------//
+// Algorithm tags.
+
+//! Neighbor operations are executed in serial on each particle thread.
+class SerialNeighborOpTag {};
+
+//! Neighbor operations are executed in parallel in a team on each particle
+//! thread.
+class TeamNeighborOpTag {};
+
+//---------------------------------------------------------------------------//
+/*!
+  \brief Execute \c functor in parallel according to the execution \c policy
+  with a thread-local serial loop over particle neighbors.
+
+  \tparam FunctorType The functor type to execute.
+
+  \tparam NeighborListType The neighbor list type.
+
+  \tparam ExecParams The Kokkos range policy parameters.
+
+  \param exec_policy The policy over which to execute the functor.
+
+  \param functor The functor to execute in parallel
+
+  \param list The neighbor list over which to execute the neighbor operations.
+
+  \param tag Algorithm tag indicating a serial loop strategy over particle
+  neighbors.
+
+  \param str An optional name for the functor. Will be forwarded to the
+  Kokkos::parallel_for called by this code and can be used for identification
+  and profiling purposes.
+
+  A "functor" is a class containing the function to execute in parallel, data
+  needed for that execution, and an optional \c execution_space typedef.  Here
+  is an example functor for neighbor parallel_for:
+
+  \code
+  class FunctorType {
+  public:
+  typedef  ...  execution_space ;
+  void operator() ( const int particle_index, const int neighbor_index ) const ;
+  };
+  \endcode
+
+  In the above example, \c Index is a Cabana index to a given AoSoA element
+  for a particle and its neighbor. Its <tt>operator()</tt> method defines the
+  operation to parallelize, over the range of indices
+  <tt>idx=[begin,end]</tt>.  This compares to a single iteration \c idx of a
+  \c for loop.
+*/
+template<class FunctorType, class NeighborListType, class ... ExecParameters>
+inline void neighbor_parallel_for(
+    const Kokkos::RangePolicy<ExecParameters...>& exec_policy,
+    const FunctorType& functor,
+    const NeighborListType& list,
+    const SerialNeighborOpTag& tag,
+    const std::string& str = "" )
+{
+    std::ignore = tag;
+
+    using work_tag =
+        typename Kokkos::RangePolicy<ExecParameters...>::work_tag;
+
+    Kokkos::parallel_for(
+        exec_policy,
+        KOKKOS_LAMBDA( const int i )
+        {
+            for ( int n = 0;
+                  n < NeighborList<NeighborListType>::numNeighbor(list,i);
+                  ++n )
+                Impl::functorTagDispatch<work_tag>(
+                    functor,
+                    i,
+                    NeighborList<NeighborListType>::getNeighbor(list,i,n) );
+        },
+        str );
+
+    Kokkos::fence();
+}
+
+//---------------------------------------------------------------------------//
+/*!
+  \brief Execute \c functor in parallel according to the execution \c policy
+  with team parallelism over particle neighbors.
+
+  \tparam ExecutionSpace The execution space in which to execute the functor.
+
+  \tparam FunctorType The functor type to execute.
+
+  \tparam NeighborListType The neighbor list type.
+
+  \param exec_policy The policy over which to execute the functor.
+
+  \param functor The functor to execute in parallel
+
+  \param list The neighbor list over which to execute the neighbor operations.
+
+  \param tag Algorithm tag indicating a team parallel strategy over particle
+  neighbors.
+
+  \param str An optional name for the functor. Will be forwarded to the
+  Kokkos::parallel_for called by this code and can be used for identification
+  and profiling purposes.
+
+  A "functor" is a class containing the function to execute in parallel, data
+  needed for that execution, and an optional \c execution_space typedef.  Here
+  is an example functor for neighbor parallel_for:
+
+  \code
+  class FunctorType {
+  public:
+  typedef  ...  execution_space ;
+  void operator() ( const int particle_index, const int neighbor_index ) const ;
+  };
+  \endcode
+
+  In the above example, \c Index is a Cabana index to a given AoSoA element
+  for a particle and its neighbor. Its <tt>operator()</tt> method defines the
+  operation to parallelize, over the range of indices
+  <tt>idx=[begin,end]</tt>.  This compares to a single iteration \c idx of a
+  \c for loop.
+*/
+template<class FunctorType, class NeighborListType, class ... ExecParameters>
+inline void neighbor_parallel_for(
+    const Kokkos::RangePolicy<ExecParameters...>& exec_policy,
+    const FunctorType& functor,
+    const NeighborListType& list,
+    const TeamNeighborOpTag& tag,
+    const std::string& str = "" )
+{
+    std::ignore = tag;
+
+    using work_tag =
+        typename Kokkos::RangePolicy<ExecParameters...>::work_tag;
+
+    using execution_space =
+        typename Kokkos::RangePolicy<ExecParameters...>::execution_space;
+
+    using kokkos_policy =
+        Kokkos::TeamPolicy<execution_space,
+                           Kokkos::Schedule<Kokkos::Dynamic> >;
+    kokkos_policy team_policy( exec_policy.end() - exec_policy.begin(),
+                               Kokkos::AUTO );
+
+    Kokkos::parallel_for(
+        team_policy,
+        KOKKOS_LAMBDA( const typename kokkos_policy::member_type& team )
+        {
+            auto i = team.league_rank() + exec_policy.begin();
+            Kokkos::parallel_for(
+                Kokkos::TeamThreadRange(
+                    team,NeighborList<NeighborListType>::numNeighbor(list,i)),
+                [&]( const int n ) {
+                Impl::functorTagDispatch<work_tag>(
+                    functor,
+                    i,
+                    NeighborList<NeighborListType>::getNeighbor(list,i,n) );
+                });
+        },
+        str );
 
     Kokkos::fence();
 }
